First draft
This is the code that I have for the classes, fragments and layouts. For the most part they're similar to the code we've already been given, the only thing is that when the application is run, while the add button will take me to fragment one, when I try to save the application crashes.